refactor: move get_tooltip/get_profile_state out of config.rs
